[FIX] 찜한 양도글 조회

+++ src/main/java/com/efub/bageasy/domain/heart/service/HeartService.java
@@ -1,11 +1,12 @@
 package com.efub.bageasy.domain.heart.service;
 
 import com.efub.bageasy.domain.heart.domain.Heart;
-import com.efub.bageasy.domain.heart.dto.HeartPostResponseDto;
 import com.efub.bageasy.domain.heart.dto.HeartResponseDto;
 import com.efub.bageasy.domain.heart.repository.HeartRepository;
+import com.efub.bageasy.domain.image.service.ImageService;
 import com.efub.bageasy.domain.member.domain.Member;
 import com.efub.bageasy.domain.post.domain.Post;
+import com.efub.bageasy.domain.post.dto.PostResponseDto;
 import com.efub.bageasy.domain.post.repository.PostRepository;
 import com.efub.bageasy.global.exception.CustomException;
 import com.efub.bageasy.global.exception.ErrorCode;
@@ -14,6 +15,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -24,10 +26,15 @@ public class HeartService {
     private final HeartRepository heartRepository;
     private final PostRepository postRepository;
 
+    private final ImageService imageService;
+
     @Transactional
     public void createByPostId(Member member, Long postId){
         postRepository.findById(postId)
                 .orElseThrow(() -> new CustomException(ErrorCode.POST_NOT_FOUND));
+        if(heartRepository.existsHeartByMemberIdAndPostId(member.getMemberId(), postId)){
+            throw new CustomException(ErrorCode.ALREADY_LIKED);
+        }
         Heart heart = new Heart(member.getMemberId(), postId);
         heartRepository.save(heart);
     }
@@ -36,6 +43,9 @@ public class HeartService {
     public void deleteByPostId(Member member, Long postId){
         postRepository.findById(postId)
                 .orElseThrow(() -> new CustomException(ErrorCode.POST_NOT_FOUND));
+        if(!heartRepository.existsHeartByMemberIdAndPostId(member.getMemberId(), postId)){
+            throw new CustomException(ErrorCode.NOT_LIKED);
+        }
         Heart heart = heartRepository.findByMemberIdAndPostId(member.getMemberId(), postId)
                 .orElseThrow(() -> new CustomException(ErrorCode.HEART_NOT_FOUND));
         heartRepository.delete(heart);
@@ -55,12 +65,12 @@ public class HeartService {
 
     //찜한 양도글 목록 조회
     @Transactional(readOnly = true)
-    public List<HeartPostResponseDto> findHeartPost(Member member){
+    public List<PostResponseDto> findHeartPost(Member member){
         return heartRepository.findByMemberId(member.getMemberId())
                 .stream()
                 .map(heart -> postRepository.findById(heart.getPostId())
                         .orElseThrow(() -> new CustomException(ErrorCode.POST_NOT_FOUND)))
-                .map(HeartPostResponseDto::new)
+                .map(post -> new PostResponseDto(post, imageService.findPostImage(post), member))
                 .collect(Collectors.toList());
     }
 }
